Fixes (GAIL, A2C and BC) + Add Pretraining (#206)
* Remove GAIL results

* Begin GAIL cleanup

* Fixes for GAIL + add tests

* Fix GAIL saving/loading

* Add notes

* Bug fixes + update changelog for GAIL

* Remove unused file

* Fix A2C with continuous actions
+ update logging

* Fix ACKTR runner

* Fixed behavior cloning + add test

* Add pretrain method

* Update coverage rc

* Rename dataset file

* Refactor expert dataset generation

* Clean up code for codacy

* Remove unused import

* Fix close method for DummyVecEnv

* Add support for pretraining with discrete actions

* Style fix

* Kill env processes to avoid memory error during tests

* Call close instead of killing processes

* Start rewriting expert dataset
In order to support image dataset

* Style fixes

* Add image dataset recorder

* Add documentation for pretraing + GAIL

* Remove pretrain with image buggy test for now

* Fix forkserver hanging when using atari env

* Reduce number of cpu for atari envs

* Change default start method (test CI)

* Test sequential mode

* Add sequential processing param

* Switch back to fork start method + add warning

* Add discrete actions support for GAIL + fix deprecations

* Update documentation

* Clean up + format code

* Replace num_iters by n_epochs + update doc

* Do not display NaN reward when not using Monitor (for SAC)

* Document Gumbel-max trick

* Simplify dataloader

* Expert model can be a callable

+++ stable_baselines/acktr/kfac.py
@@ -469,11 +469,11 @@ class KfacOptimizer:
 
                     # assume sampled loss is averaged. TODO:figure out better
                     # way to handle this
-                    bprop_factor *= tf.to_float(batch_size)
+                    bprop_factor *= tf.cast(batch_size, tf.float32)
                     ##
 
-                    cov_b = tf.matmul(
-                        bprop_factor, bprop_factor, transpose_a=True) / tf.to_float(tf.shape(bprop_factor)[0])
+                    cov_b = tf.matmul(bprop_factor, bprop_factor,
+                                      transpose_a=True) / tf.cast(tf.shape(bprop_factor)[0], tf.float32)
 
                     update_ops.append(cov_b)
                     stats_updates[stats_var] = cov_b
