+++ stable_baselines/common/distributions.py
@@ -293,7 +293,6 @@ class CategoricalProbabilityDistribution(ProbabilityDistribution):
         return tf.argmax(self.logits, axis=-1)
 
     def neglogp(self, x):
-        # return tf.nn. (logits=self.logits, labels=x)
         # Note: we can't use sparse_softmax_cross_entropy_with_logits because
         #       the implementation does not allow second-order derivatives...
         one_hot_actions = tf.one_hot(x, self.logits.get_shape().as_list()[-1])
@@ -319,6 +318,8 @@ class CategoricalProbabilityDistribution(ProbabilityDistribution):
         return tf.reduce_sum(p_0 * (tf.log(z_0) - a_0), axis=-1)
 
     def sample(self):
+        # Gumbel-max trick to sample
+        # a categorical distribution (see http://amid.fish/humble-gumbel)
         uniform = tf.random_uniform(tf.shape(self.logits), dtype=self.logits.dtype)
         return tf.argmax(self.logits - tf.log(-tf.log(uniform)), axis=-1)
 
@@ -395,7 +396,7 @@ class DiagGaussianProbabilityDistribution(ProbabilityDistribution):
 
     def neglogp(self, x):
         return 0.5 * tf.reduce_sum(tf.square((x - self.mean) / self.std), axis=-1) \
-               + 0.5 * np.log(2.0 * np.pi) * tf.to_float(tf.shape(x)[-1]) \
+               + 0.5 * np.log(2.0 * np.pi) * tf.cast(tf.shape(x)[-1], tf.float32) \
                + tf.reduce_sum(self.logstd, axis=-1)
 
     def kl(self, other):
@@ -439,7 +440,8 @@ class BernoulliProbabilityDistribution(ProbabilityDistribution):
         return tf.round(self.probabilities)
 
     def neglogp(self, x):
-        return tf.reduce_sum(tf.nn.sigmoid_cross_entropy_with_logits(logits=self.logits, labels=tf.to_float(x)),
+        return tf.reduce_sum(tf.nn.sigmoid_cross_entropy_with_logits(logits=self.logits,
+                                                                     labels=tf.cast(x, tf.float32)),
                              axis=-1)
 
     def kl(self, other):
@@ -454,7 +456,7 @@ class BernoulliProbabilityDistribution(ProbabilityDistribution):
 
     def sample(self):
         samples_from_uniform = tf.random_uniform(tf.shape(self.probabilities))
-        return tf.to_float(math_ops.less(samples_from_uniform, self.probabilities))
+        return tf.cast(math_ops.less(samples_from_uniform, self.probabilities), tf.float32)
 
     @classmethod
     def fromflat(cls, flat):
