+++ stable_baselines/common/mpi_running_mean_std.py
@@ -31,8 +31,9 @@ class RunningMeanStd(object):
             name="count", trainable=False)
         self.shape = shape
 
-        self.mean = tf.to_float(self._sum / self._count)
-        self.std = tf.sqrt(tf.maximum(tf.to_float(self._sumsq / self._count) - tf.square(self.mean), 1e-2))
+        self.mean = tf.cast(self._sum / self._count, tf.float32)
+        self.std = tf.sqrt(tf.maximum(tf.cast(self._sumsq / self._count, tf.float32) - tf.square(self.mean),
+                                      1e-2))
 
         newsum = tf.placeholder(shape=self.shape, dtype=tf.float64, name='sum')
         newsumsq = tf.placeholder(shape=self.shape, dtype=tf.float64, name='var')
